refactor: extract funcs to credentials package
